move StringParam => proptest::string
