+++ src/arbitrary/_std/mod.rs
@@ -38,7 +38,6 @@ mod rc;
 mod result;
 mod str;
 mod string;
-pub use self::string::*;
 mod sync;
 mod thread;
 mod time;
